fully-fleshed out set-type voicing listener. Best used with keyboard macros that should receive e-lisp functions to call instead.
should now be re-factored to strip out "magic numbers" and allow for
greater flexibilty in parameter setting.

 # Please enter the commit message for your changes. Lines
starting # with '#' will be ignored, and an empty message aborts the
commit. # # Committer: a <a@as-MacBook-Pro-2.local> # # On branch
master # Changes to be committed: # (use "git reset HEAD <file>..." to
unstage) # # modified: src/tutorial/core.clj #